re-add accidentally-deleted documentation
