+++ EcoSoapBank/Helpers/UIKit/UIViews+Convenience.swift
@@ -1,0 +1,51 @@
+//
+//  UIElementInitializers.swift
+//  EcoSoapBank
+//
+//  Created by Christopher Devito on 9/1/20.
+//  Copyright © 2020 Spencer Curtis. All rights reserved.
+//
+
+import UIKit
+
+extension UILabel {
+    /**
+     Initializes a UILabel with title, frame,alignment, and textColor passed in.
+     - Parameter title: **String** that describes the UILabel's text property
+     - Parameter frame: **CGRect** that describes the initial dimensions of the UILabel
+     - Parameter alignment: **NSTextAlignment** that describes the UILabel's  textAlignment property
+     - Parameter textColor: **UIColor** that describes the UILabel's textColor property
+     */
+    convenience init(_ title: String, frame: CGRect, alignment: NSTextAlignment, textColor: UIColor = .label) {
+        self.init(frame: frame)
+        translatesAutoresizingMaskIntoConstraints = false
+        text = title
+        textAlignment = alignment
+        self.textColor = textColor
+    }
+}
+
+extension UIImageView {
+    /// Initializes a UIImageView with image name
+    /// - Parameter systemName: **String** with the system name for the image
+    convenience init(_ systemName: String) {
+        self.init(image: UIImage(systemName: systemName, withConfiguration: UIImage.SymbolConfiguration(pointSize: 16, weight: .semibold)))
+        translatesAutoresizingMaskIntoConstraints = false
+        tintColor = .lightGray
+    }
+}
+
+extension UIStackView {
+    convenience init(
+        axis: NSLayoutConstraint.Axis = .vertical,
+        alignment: UIStackView.Alignment = .fill,
+        distribution: UIStackView.Distribution = .fill,
+        spacing: CGFloat = 8
+    ) {
+        self.init()
+        self.axis = axis
+        self.alignment = alignment
+        self.distribution = distribution
+        self.spacing = spacing
+    }
+}
